fix: changed long double to double, add xor keyword

+++ lib/lexer/tokens/TokenFactory.cpp
@@ -52,7 +52,7 @@ std::unique_ptr<Token> MakeIntLiteral(std::string raw, int64_t v, int32_t line, 
   return std::make_unique<LiteralToken>(std::move(raw), std::make_unique<IntValue>(v), TokenPosition(line, col));
 }
 
-std::unique_ptr<Token> MakeFloatLiteral(std::string raw, long double v, int32_t line, int32_t col) {
+std::unique_ptr<Token> MakeFloatLiteral(std::string raw, double v, int32_t line, int32_t col) {
   return std::make_unique<LiteralToken>(std::move(raw), std::make_unique<FloatValue>(v), TokenPosition(line, col));
 }
